Mark empty files as being specifically empty by design.
When serializing queries, remember that arrays in Lua start at 1,
so no need to add 1 to the indices.

The member rules are in rules.member, not rules.members.

If a member has a different name, it's called locationName, not
name.

Finally, the split routine sometimes leaves a blank entry on the
end.  Apparently some consumers of the library expect that, but when
serializing a list with a custom name it doesn't work right, so look
for that and remove it.

This makes the sendMessageBatch and deleteMessageBatch routines work
properly for SQS.